Code/Processing/requirements.txt
demucs>=4.0.0
pyannote.audio>=3.1.0
librosa>=0.10.0
soundfile>=0.12.0
tqdm>=4.64.0
pandas>=1.5.0
numpy>=1.21.0
torch>=1.12.0
torchaudio>=0.12.0
